+++ src/app/Http/Controllers/Operations/CloneOperation.php
@@ -29,7 +29,7 @@ trait CloneOperation
     {
         $this->crud->allowAccess('clone');
 
-        $this->crud->operation('list', function () {
+        $this->crud->operation(['list', 'show'], function () {
             $this->crud->addButton('line', 'clone', 'view', 'crud::buttons.clone', 'end');
         });
     }