+++ src/resources/views/show.blade.php
@@ -72,7 +72,7 @@
                         </td>
 		            </tr>
 		        @endforeach
-				@if ($crud->buttons->where('stack', 'line')->count())
+				@if ($crud->buttons()->where('stack', 'line')->count())
 					<tr>
 						<td><strong>{{ trans('backpack::crud.actions') }}</strong></td>
 						<td>